Guzzless built. Preparation to tests.
