+++ .travis.yml
@@ -1,9 +1,8 @@
 language: php
-before_install: bash bin/build_etcd
 php:
 - 5.4
 - 5.5
-- hhvm
+- 7.0
 before_script:
 - composer install
 script: vendor/bin/phpunit
