+++ composer.json
@@ -16,7 +16,7 @@
         "guzzlehttp/guzzle": "^6.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "~5.0"
     },
     "autoload": {
         "psr-4": {
